+++ drivers/staging/vt6655/key.c
@@ -135,7 +135,7 @@ void KeyvInitTable (PSKeyManagement pTable, unsigned long dwIoBase)
 BOOL KeybGetKey (
     PSKeyManagement pTable,
     unsigned char *pbyBSSID,
-    DWORD           dwKeyIndex,
+    unsigned long dwKeyIndex,
     PSKeyItem       *pKey
     )
 {
@@ -193,7 +193,7 @@ BOOL KeybGetKey (
 BOOL KeybSetKey (
     PSKeyManagement pTable,
     unsigned char *pbyBSSID,
-    DWORD           dwKeyIndex,
+    unsigned long dwKeyIndex,
     unsigned long uKeyLength,
     PQWORD          pKeyRSC,
     unsigned char *pbyKey,
@@ -366,7 +366,7 @@ BOOL KeybSetKey (
 BOOL KeybRemoveKey (
     PSKeyManagement pTable,
     unsigned char *pbyBSSID,
-    DWORD           dwKeyIndex,
+    unsigned long dwKeyIndex,
     unsigned long dwIoBase
     )
 {
@@ -473,7 +473,7 @@ BOOL KeybRemoveAllKey (
  */
 void KeyvRemoveWEPKey (
     PSKeyManagement pTable,
-    DWORD           dwKeyIndex,
+    unsigned long dwKeyIndex,
     unsigned long dwIoBase
     )
 {
@@ -521,7 +521,7 @@ void KeyvRemoveAllWEPKey (
 BOOL KeybGetTransmitKey (
     PSKeyManagement pTable,
     unsigned char *pbyBSSID,
-    DWORD           dwKeyType,
+    unsigned long dwKeyType,
     PSKeyItem       *pKey
     )
 {
@@ -634,7 +634,7 @@ BOOL KeybCheckPairewiseKey (
  */
 BOOL KeybSetDefaultKey (
     PSKeyManagement pTable,
-    DWORD           dwKeyIndex,
+    unsigned long dwKeyIndex,
     unsigned long uKeyLength,
     PQWORD          pKeyRSC,
     unsigned char *pbyKey,
@@ -741,7 +741,7 @@ BOOL KeybSetDefaultKey (
  */
 BOOL KeybSetAllGroupKey (
     PSKeyManagement pTable,
-    DWORD           dwKeyIndex,
+    unsigned long dwKeyIndex,
     unsigned long uKeyLength,
     PQWORD          pKeyRSC,
     unsigned char *pbyKey,
