+++ drivers/staging/vt6655/mac.c
@@ -187,7 +187,7 @@ BOOL MACbIsRegBitsOff (unsigned long dwIoBase, BYTE byRegOfs, BYTE byTestBits)
  */
 BOOL MACbIsIntDisable (unsigned long dwIoBase)
 {
-    DWORD dwData;
+    unsigned long dwData;
 
     VNSvInPortD(dwIoBase + MAC_REG_IMR, &dwData);
     if (dwData != 0)
@@ -596,7 +596,7 @@ void MACvSetPacketFilter (unsigned long dwIoBase, WORD wFilterType)
         // set multicast address to accept none
         MACvSelectPage1(dwIoBase);
         VNSvOutPortD(dwIoBase + MAC_REG_MAR0, 0L);
-        VNSvOutPortD(dwIoBase + MAC_REG_MAR0 + sizeof(DWORD), 0L);
+        VNSvOutPortD(dwIoBase + MAC_REG_MAR0 + sizeof(unsigned long), 0L);
         MACvSelectPage0(dwIoBase);
     }
 
@@ -604,7 +604,7 @@ void MACvSetPacketFilter (unsigned long dwIoBase, WORD wFilterType)
         // set multicast address to accept all
         MACvSelectPage1(dwIoBase);
         VNSvOutPortD(dwIoBase + MAC_REG_MAR0, 0xFFFFFFFFL);
-        VNSvOutPortD(dwIoBase + MAC_REG_MAR0 + sizeof(DWORD), 0xFFFFFFFFL);
+        VNSvOutPortD(dwIoBase + MAC_REG_MAR0 + sizeof(unsigned long), 0xFFFFFFFFL);
         MACvSelectPage0(dwIoBase);
     }
 
@@ -731,7 +731,7 @@ void MACvRestoreContext (unsigned long dwIoBase, unsigned char *pbyCxtBuf)
  */
 BOOL MACbCompareContext (unsigned long dwIoBase, unsigned char *pbyCxtBuf)
 {
-    DWORD       dwData;
+    unsigned long dwData;
 
     // compare MAC context to determine if this is a power lost init,
     // return TRUE for power remaining init, return FALSE for power lost init
@@ -843,7 +843,7 @@ BOOL MACbSafeSoftwareReset (unsigned long dwIoBase)
 BOOL MACbSafeRxOff (unsigned long dwIoBase)
 {
     WORD    ww;
-    DWORD   dwData;
+    unsigned long dwData;
     BYTE    byData;
 
     // turn off wow temp for turn off Rx safely
@@ -904,7 +904,7 @@ BOOL MACbSafeRxOff (unsigned long dwIoBase)
 BOOL MACbSafeTxOff (unsigned long dwIoBase)
 {
     WORD    ww;
-    DWORD   dwData;
+    unsigned long dwData;
     BYTE    byData;
 
     // Clear TX DMA
@@ -1080,7 +1080,7 @@ void MACvInitialize (unsigned long dwIoBase)
  * Return Value: none
  *
  */
-void MACvSetCurrRx0DescAddr (unsigned long dwIoBase, DWORD dwCurrDescAddr)
+void MACvSetCurrRx0DescAddr (unsigned long dwIoBase, unsigned long dwCurrDescAddr)
 {
 WORD    ww;
 BYTE    byData;
@@ -1118,7 +1118,7 @@ BYTE    byOrgDMACtl;
  * Return Value: none
  *
  */
-void MACvSetCurrRx1DescAddr (unsigned long dwIoBase, DWORD dwCurrDescAddr)
+void MACvSetCurrRx1DescAddr (unsigned long dwIoBase, unsigned long dwCurrDescAddr)
 {
 WORD    ww;
 BYTE    byData;
@@ -1156,7 +1156,7 @@ BYTE    byOrgDMACtl;
  * Return Value: none
  *
  */
-void MACvSetCurrTx0DescAddrEx (unsigned long dwIoBase, DWORD dwCurrDescAddr)
+void MACvSetCurrTx0DescAddrEx (unsigned long dwIoBase, unsigned long dwCurrDescAddr)
 {
 WORD    ww;
 BYTE    byData;
@@ -1195,7 +1195,7 @@ BYTE    byOrgDMACtl;
  *
  */
  //TxDMA1 = AC0DMA
-void MACvSetCurrAC0DescAddrEx (unsigned long dwIoBase, DWORD dwCurrDescAddr)
+void MACvSetCurrAC0DescAddrEx (unsigned long dwIoBase, unsigned long dwCurrDescAddr)
 {
 WORD    ww;
 BYTE    byData;
@@ -1222,7 +1222,7 @@ BYTE    byOrgDMACtl;
 
 
 
-void MACvSetCurrTXDescAddr (int iTxType, unsigned long dwIoBase, DWORD dwCurrDescAddr)
+void MACvSetCurrTXDescAddr (int iTxType, unsigned long dwIoBase, unsigned long dwCurrDescAddr)
 {
     if(iTxType == TYPE_AC0DMA){
         MACvSetCurrAC0DescAddrEx(dwIoBase, dwCurrDescAddr);
@@ -1310,7 +1310,7 @@ void MACvOneShotTimer1MicroSec (unsigned long dwIoBase, unsigned int uDelayTime)
 }
 
 
-void MACvSetMISCFifo (unsigned long dwIoBase, WORD wOffset, DWORD dwData)
+void MACvSetMISCFifo (unsigned long dwIoBase, WORD wOffset, unsigned long dwData)
 {
     if (wOffset > 273)
         return;
@@ -1350,7 +1350,7 @@ unsigned int ww = 0;
 
 void MACvClearBusSusInd (unsigned long dwIoBase)
 {
-    DWORD dwOrgValue;
+    unsigned long dwOrgValue;
     unsigned int ww;
     // check if BcnSusInd enabled
     VNSvInPortD(dwIoBase + MAC_REG_ENCFG , &dwOrgValue);
@@ -1373,7 +1373,7 @@ void MACvClearBusSusInd (unsigned long dwIoBase)
 void MACvEnableBusSusEn (unsigned long dwIoBase)
 {
     BYTE  byOrgValue;
-    DWORD dwOrgValue;
+    unsigned long dwOrgValue;
     unsigned int ww;
     // check if BcnSusInd enabled
     VNSvInPortB(dwIoBase + MAC_REG_CFG , &byOrgValue);
@@ -1460,7 +1460,7 @@ void MACvSetKeyEntry (unsigned long dwIoBase, WORD wKeyCtl, unsigned int uEntryI
 		unsigned int uKeyIdx, unsigned char *pbyAddr, unsigned long *pdwKey, BYTE byLocalID)
 {
 WORD    wOffset;
-DWORD   dwData;
+unsigned long dwData;
 int     ii;
 
     if (byLocalID <= 1)
@@ -1555,7 +1555,7 @@ void MACvSetDefaultKeyEntry (unsigned long dwIoBase, unsigned int uKeyLen,
 		unsigned int uKeyIdx, unsigned long *pdwKey, BYTE byLocalID)
 {
 WORD    wOffset;
-DWORD   dwData;
+unsigned long dwData;
 int     ii;
 
     if (byLocalID <= 1)
@@ -1605,7 +1605,7 @@ int     ii;
 void MACvEnableDefaultKey (unsigned long dwIoBase, BYTE byLocalID)
 {
 WORD    wOffset;
-DWORD   dwData;
+unsigned long dwData;
 
 
     if (byLocalID <= 1)
@@ -1640,7 +1640,7 @@ DWORD   dwData;
 void MACvDisableDefaultKey (unsigned long dwIoBase)
 {
 WORD    wOffset;
-DWORD   dwData;
+unsigned long dwData;
 
 
     wOffset = MISCFIFO_KEYETRY0;
@@ -1671,7 +1671,7 @@ void MACvSetDefaultTKIPKeyEntry (unsigned long dwIoBase, unsigned int uKeyLen,
 		unsigned int uKeyIdx, unsigned long *pdwKey, BYTE byLocalID)
 {
 WORD    wOffset;
-DWORD   dwData;
+unsigned long dwData;
 int     ii;
 
     if (byLocalID <= 1)
@@ -1727,7 +1727,7 @@ int     ii;
 void MACvSetDefaultKeyCtl (unsigned long dwIoBase, WORD wKeyCtl, unsigned int uEntryIdx, BYTE byLocalID)
 {
 WORD    wOffset;
-DWORD   dwData;
+unsigned long dwData;
 
     if (byLocalID <= 1)
         return;
