+++ drivers/staging/vt6655/rxtx.h
@@ -43,7 +43,7 @@ FASTVC-BENCH-CORPUS:linux-main-100k-v1-e0bd41dc-8e12-4f1b-978d-a3645ae59da0
 /*
 void vGenerateMACHeader(
     PSDevice pDevice,
-    DWORD dwTxBufferAddr,
+    unsigned long dwTxBufferAddr,
     unsigned char *pbySkbData,
     unsigned int cbPacketSize,
     BOOL bDMA0Used,
@@ -53,7 +53,7 @@ void vGenerateMACHeader(
 
 void vProcessRxMACHeader (
     PSDevice pDevice,
-    DWORD dwRxBufferAddr,
+    unsigned long dwRxBufferAddr,
     unsigned int cbPacketSize,
     BOOL bIsWEP,
     PUINT pcbHeadSize
